% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_dist_plot_directed_function.R
\name{deg_dist_plot_directed}
\alias{deg_dist_plot_directed}
\title{deg_dist_plot_directed}
\usage{
deg_dist_plot_directed(NetList, observed_name)
}
\arguments{
\item{NetList}{Network list created using read_sim_networks function}

\item{observed_name}{Observed network .txt file (used in PNet)}
}
\value{
degree distribution plots
}
\description{
Creates degree distribution plots for directed networks
}
\examples{
\donttest{
degplots<-deg_dist_plot_directed(NetList,"Oberved Directed Network.txt")
}
}
