% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sim_networks_function.R
\name{read_sim_networks}
\alias{read_sim_networks}
\title{read_sim_networks}
\usage{
read_sim_networks(searchname)
}
\arguments{
\item{searchname}{Search term - term to extract all simulated networks.}
}
\value{
network list
}
\description{
Reads in all simulated networks (.txt files) into R as network files from the working directory.
}
\examples{
\donttest{
net_list<-read_sim_networks("PNET_Network_A_")
}
}
