% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPNetStatistics_function.R
\name{readPNetStatistics}
\alias{readPNetStatistics}
\title{readPNetStatistics}
\usage{
readPNetStatistics(filename)
}
\arguments{
\item{filename}{Simulated .txt file}
}
\value{
matrix
}
\description{
Extracts matrix from PNet simulated text file. Taken from MPNet manual.
}
\examples{
\donttest{
SimADJ <- readPNetStatistics('mytest_Network_A_1001000.txt')
}
}
